Python examples: Added a minimal 'getting started'
This adds a simple README with (1) an description of the examples in this directory, and (2) the minimal steps to getting started. 

Since this page is the top result when people search "Fission Python", this README should help recurring users to quickly find what they are looking for and point new users to the official docs.